SVM: Add doc comments, restrict visibility of some xfaces to crate
