+++ svm/src/account_loader.rs
@@ -38,8 +38,11 @@ use {
 };
 
 // for the load instructions
-pub type TransactionRent = u64;
-pub type TransactionProgramIndices = Vec<Vec<IndexOfAccount>>;
+pub(crate) type TransactionRent = u64;
+pub(crate) type TransactionProgramIndices = Vec<Vec<IndexOfAccount>>;
+pub type TransactionCheckResult = (transaction::Result<()>, Option<NoncePartial>, Option<u64>);
+pub type TransactionLoadResult = (Result<LoadedTransaction>, Option<NonceFull>);
+
 #[derive(PartialEq, Eq, Debug, Clone)]
 pub struct LoadedTransaction {
     pub accounts: Vec<TransactionAccount>,
@@ -48,10 +51,66 @@ pub struct LoadedTransaction {
     pub rent_debits: RentDebits,
 }
 
-pub type TransactionLoadResult = (Result<LoadedTransaction>, Option<NonceFull>);
-pub type TransactionCheckResult = (transaction::Result<()>, Option<NoncePartial>, Option<u64>);
+/// Check whether the payer_account is capable of paying the fee. The
+/// side effect is to subtract the fee amount from the payer_account
+/// balance of lamports. If the payer_acount is not able to pay the
+/// fee, the error_counters is incremented, and a specific error is
+/// returned.
+pub fn validate_fee_payer(
+    payer_address: &Pubkey,
+    payer_account: &mut AccountSharedData,
+    payer_index: IndexOfAccount,
+    error_counters: &mut TransactionErrorMetrics,
+    rent_collector: &RentCollector,
+    fee: u64,
+) -> Result<()> {
+    if payer_account.lamports() == 0 {
+        error_counters.account_not_found += 1;
+        return Err(TransactionError::AccountNotFound);
+    }
+    let system_account_kind = get_system_account_kind(payer_account).ok_or_else(|| {
+        error_counters.invalid_account_for_fee += 1;
+        TransactionError::InvalidAccountForFee
+    })?;
+    let min_balance = match system_account_kind {
+        SystemAccountKind::System => 0,
+        SystemAccountKind::Nonce => {
+            // Should we ever allow a fees charge to zero a nonce account's
+            // balance. The state MUST be set to uninitialized in that case
+            rent_collector.rent.minimum_balance(NonceState::size())
+        }
+    };
 
-pub fn load_accounts<CB: TransactionProcessingCallback>(
+    payer_account
+        .lamports()
+        .checked_sub(min_balance)
+        .and_then(|v| v.checked_sub(fee))
+        .ok_or_else(|| {
+            error_counters.insufficient_funds += 1;
+            TransactionError::InsufficientFundsForFee
+        })?;
+
+    let payer_pre_rent_state = RentState::from_account(payer_account, &rent_collector.rent);
+    payer_account
+        .checked_sub_lamports(fee)
+        .map_err(|_| TransactionError::InsufficientFundsForFee)?;
+
+    let payer_post_rent_state = RentState::from_account(payer_account, &rent_collector.rent);
+    RentState::check_rent_state_with_account(
+        &payer_pre_rent_state,
+        &payer_post_rent_state,
+        payer_address,
+        payer_account,
+        payer_index,
+    )
+}
+
+/// Collect information about accounts used in txs transactions and
+/// return vector of tuples, one for each transaction in the
+/// batch. Each tuple contains struct of information about accounts as
+/// its first element and an optional transaction nonce info as its
+/// second element.
+pub(crate) fn load_accounts<CB: TransactionProcessingCallback>(
     callbacks: &CB,
     txs: &[SanitizedTransaction],
     lock_results: &[TransactionCheckResult],
@@ -399,55 +458,6 @@ fn accumulate_and_check_loaded_account_data_size(
     }
 }
 
-pub fn validate_fee_payer(
-    payer_address: &Pubkey,
-    payer_account: &mut AccountSharedData,
-    payer_index: IndexOfAccount,
-    error_counters: &mut TransactionErrorMetrics,
-    rent_collector: &RentCollector,
-    fee: u64,
-) -> Result<()> {
-    if payer_account.lamports() == 0 {
-        error_counters.account_not_found += 1;
-        return Err(TransactionError::AccountNotFound);
-    }
-    let system_account_kind = get_system_account_kind(payer_account).ok_or_else(|| {
-        error_counters.invalid_account_for_fee += 1;
-        TransactionError::InvalidAccountForFee
-    })?;
-    let min_balance = match system_account_kind {
-        SystemAccountKind::System => 0,
-        SystemAccountKind::Nonce => {
-            // Should we ever allow a fees charge to zero a nonce account's
-            // balance. The state MUST be set to uninitialized in that case
-            rent_collector.rent.minimum_balance(NonceState::size())
-        }
-    };
-
-    payer_account
-        .lamports()
-        .checked_sub(min_balance)
-        .and_then(|v| v.checked_sub(fee))
-        .ok_or_else(|| {
-            error_counters.insufficient_funds += 1;
-            TransactionError::InsufficientFundsForFee
-        })?;
-
-    let payer_pre_rent_state = RentState::from_account(payer_account, &rent_collector.rent);
-    payer_account
-        .checked_sub_lamports(fee)
-        .map_err(|_| TransactionError::InsufficientFundsForFee)?;
-
-    let payer_post_rent_state = RentState::from_account(payer_account, &rent_collector.rent);
-    RentState::check_rent_state_with_account(
-        &payer_pre_rent_state,
-        &payer_post_rent_state,
-        payer_address,
-        payer_account,
-        payer_index,
-    )
-}
-
 fn construct_instructions_account(message: &SanitizedMessage) -> AccountSharedData {
     AccountSharedData::from(Account {
         data: construct_instructions_data(&message.decompile_instructions()),
@@ -460,11 +470,15 @@ fn construct_instructions_account(message: &SanitizedMessage) -> AccountSharedDa
 mod tests {
     use {
         super::*,
-        crate::transaction_processor::TransactionProcessingCallback,
+        crate::{
+            transaction_account_state_info::TransactionAccountStateInfo,
+            transaction_processor::TransactionProcessingCallback,
+        },
         nonce::state::Versions as NonceVersions,
         solana_program_runtime::{
+            compute_budget::ComputeBudget,
             compute_budget_processor,
-            loaded_programs::LoadedProgram,
+            loaded_programs::{LoadedProgram, LoadedProgramsForTxBatch},
             prioritization_fee::{PrioritizationFeeDetails, PrioritizationFeeType},
         },
         solana_sdk::{
@@ -473,22 +487,27 @@ mod tests {
             compute_budget::ComputeBudgetInstruction,
             epoch_schedule::EpochSchedule,
             feature_set::FeatureSet,
+            fee::FeeStructure,
             hash::Hash,
             instruction::CompiledInstruction,
             message::{
                 v0::{LoadedAddresses, LoadedMessage},
                 LegacyMessage, Message, MessageHeader, SanitizedMessage,
             },
+            native_loader,
+            native_token::sol_to_lamports,
             nonce,
+            nonce_info::{NonceFull, NoncePartial},
             pubkey::Pubkey,
             rent::Rent,
-            rent_collector::RentCollector,
+            rent_collector::{RentCollector, RENT_EXEMPT_RENT_EPOCH},
+            rent_debits::RentDebits,
             signature::{Keypair, Signature, Signer},
-            system_program, sysvar,
-            transaction::{Result, Transaction, TransactionError},
-            transaction_context::TransactionAccount,
+            system_program, system_transaction, sysvar,
+            transaction::{Result, SanitizedTransaction, Transaction, TransactionError},
+            transaction_context::{TransactionAccount, TransactionContext},
         },
-        std::{borrow::Cow, convert::TryFrom, sync::Arc},
+        std::{borrow::Cow, collections::HashMap, convert::TryFrom, sync::Arc},
     };
 
     #[derive(Default)]
@@ -2015,6 +2034,250 @@ mod tests {
                 rent: 0,
                 rent_debits: RentDebits::default()
             }
+        );
+    }
+
+    #[test]
+    fn test_rent_state_list_len() {
+        let mint_keypair = Keypair::new();
+        let mut bank = TestCallbacks::default();
+        let recipient = Pubkey::new_unique();
+        let last_block_hash = Hash::new_unique();
+
+        let mut system_data = AccountSharedData::default();
+        system_data.set_executable(true);
+        system_data.set_owner(native_loader::id());
+        bank.accounts_map
+            .insert(Pubkey::new_from_array([0u8; 32]), system_data);
+
+        let mut mint_data = AccountSharedData::default();
+        mint_data.set_lamports(2);
+        bank.accounts_map.insert(mint_keypair.pubkey(), mint_data);
+
+        bank.accounts_map
+            .insert(recipient, AccountSharedData::default());
+
+        let tx = system_transaction::transfer(
+            &mint_keypair,
+            &recipient,
+            sol_to_lamports(1.),
+            last_block_hash,
+        );
+        let num_accounts = tx.message().account_keys.len();
+        let sanitized_tx = SanitizedTransaction::from_transaction_for_tests(tx);
+        let mut error_counters = TransactionErrorMetrics::default();
+        let loaded_txs = load_accounts(
+            &bank,
+            &[sanitized_tx.clone()],
+            &[(Ok(()), None, Some(0))],
+            &mut error_counters,
+            &FeeStructure::default(),
+            None,
+            &HashMap::new(),
+            &LoadedProgramsForTxBatch::default(),
+        );
+
+        let compute_budget = ComputeBudget::new(u64::from(
+            compute_budget_processor::DEFAULT_INSTRUCTION_COMPUTE_UNIT_LIMIT,
+        ));
+        let transaction_context = TransactionContext::new(
+            loaded_txs[0].0.as_ref().unwrap().accounts.clone(),
+            Rent::default(),
+            compute_budget.max_invoke_stack_height,
+            compute_budget.max_instruction_trace_length,
+        );
+
+        assert_eq!(
+            TransactionAccountStateInfo::new(
+                &Rent::default(),
+                &transaction_context,
+                sanitized_tx.message()
+            )
+            .len(),
+            num_accounts,
+        );
+    }
+
+    #[test]
+    fn test_load_accounts_success() {
+        let key1 = Keypair::new();
+        let key2 = Keypair::new();
+        let key3 = Keypair::new();
+        let key4 = Keypair::new();
+
+        let message = Message {
+            account_keys: vec![key2.pubkey(), key1.pubkey(), key4.pubkey()],
+            header: MessageHeader::default(),
+            instructions: vec![
+                CompiledInstruction {
+                    program_id_index: 1,
+                    accounts: vec![0],
+                    data: vec![],
+                },
+                CompiledInstruction {
+                    program_id_index: 1,
+                    accounts: vec![2],
+                    data: vec![],
+                },
+            ],
+            recent_blockhash: Hash::default(),
+        };
+
+        let legacy = LegacyMessage::new(message);
+        let sanitized_message = SanitizedMessage::Legacy(legacy);
+        let mut mock_bank = TestCallbacks::default();
+        let mut account_data = AccountSharedData::default();
+        account_data.set_executable(true);
+        account_data.set_owner(key3.pubkey());
+        mock_bank.accounts_map.insert(key1.pubkey(), account_data);
+
+        let mut account_data = AccountSharedData::default();
+        account_data.set_lamports(200);
+        mock_bank.accounts_map.insert(key2.pubkey(), account_data);
+
+        let mut account_data = AccountSharedData::default();
+        account_data.set_executable(true);
+        account_data.set_owner(native_loader::id());
+        mock_bank.accounts_map.insert(key3.pubkey(), account_data);
+
+        let mut error_counter = TransactionErrorMetrics::default();
+        let loaded_programs = LoadedProgramsForTxBatch::default();
+
+        let sanitized_transaction = SanitizedTransaction::new_for_tests(
+            sanitized_message,
+            vec![Signature::new_unique()],
+            false,
+        );
+        let lock_results =
+            (Ok(()), Some(NoncePartial::default()), Some(20u64)) as TransactionCheckResult;
+
+        let results = load_accounts(
+            &mock_bank,
+            &[sanitized_transaction],
+            &[lock_results],
+            &mut error_counter,
+            &FeeStructure::default(),
+            None,
+            &HashMap::new(),
+            &loaded_programs,
+        );
+
+        let mut account_data = AccountSharedData::default();
+        account_data.set_rent_epoch(RENT_EXEMPT_RENT_EPOCH);
+
+        assert_eq!(results.len(), 1);
+        let (loaded_result, nonce) = results[0].clone();
+        assert_eq!(
+            loaded_result.unwrap(),
+            LoadedTransaction {
+                accounts: vec![
+                    (
+                        key2.pubkey(),
+                        mock_bank.accounts_map[&key2.pubkey()].clone()
+                    ),
+                    (
+                        key1.pubkey(),
+                        mock_bank.accounts_map[&key1.pubkey()].clone()
+                    ),
+                    (key4.pubkey(), account_data),
+                    (
+                        key3.pubkey(),
+                        mock_bank.accounts_map[&key3.pubkey()].clone()
+                    ),
+                ],
+                program_indices: vec![vec![3, 1], vec![3, 1]],
+                rent: 0,
+                rent_debits: RentDebits::default()
+            }
+        );
+
+        assert_eq!(
+            nonce.unwrap(),
+            NonceFull::new(
+                Pubkey::from([0; 32]),
+                AccountSharedData::default(),
+                Some(mock_bank.accounts_map[&key2.pubkey()].clone())
+            )
+        );
+    }
+
+    #[test]
+    fn test_load_accounts_error() {
+        let mock_bank = TestCallbacks::default();
+        let message = Message {
+            account_keys: vec![Pubkey::new_from_array([0; 32])],
+            header: MessageHeader::default(),
+            instructions: vec![CompiledInstruction {
+                program_id_index: 0,
+                accounts: vec![],
+                data: vec![],
+            }],
+            recent_blockhash: Hash::default(),
+        };
+
+        let legacy = LegacyMessage::new(message);
+        let sanitized_message = SanitizedMessage::Legacy(legacy);
+        let sanitized_transaction = SanitizedTransaction::new_for_tests(
+            sanitized_message,
+            vec![Signature::new_unique()],
+            false,
+        );
+
+        let lock_results = (Ok(()), Some(NoncePartial::default()), None) as TransactionCheckResult;
+        let fee_structure = FeeStructure::default();
+
+        let result = load_accounts(
+            &mock_bank,
+            &[sanitized_transaction.clone()],
+            &[lock_results],
+            &mut TransactionErrorMetrics::default(),
+            &fee_structure,
+            None,
+            &HashMap::new(),
+            &LoadedProgramsForTxBatch::default(),
+        );
+
+        assert_eq!(
+            result,
+            vec![(Err(TransactionError::BlockhashNotFound), None)]
+        );
+
+        let lock_results =
+            (Ok(()), Some(NoncePartial::default()), Some(20u64)) as TransactionCheckResult;
+
+        let result = load_accounts(
+            &mock_bank,
+            &[sanitized_transaction.clone()],
+            &[lock_results.clone()],
+            &mut TransactionErrorMetrics::default(),
+            &fee_structure,
+            None,
+            &HashMap::new(),
+            &LoadedProgramsForTxBatch::default(),
+        );
+
+        assert_eq!(result, vec![(Err(TransactionError::AccountNotFound), None)]);
+
+        let lock_results = (
+            Err(TransactionError::InvalidWritableAccount),
+            Some(NoncePartial::default()),
+            Some(20u64),
+        ) as TransactionCheckResult;
+
+        let result = load_accounts(
+            &mock_bank,
+            &[sanitized_transaction.clone()],
+            &[lock_results],
+            &mut TransactionErrorMetrics::default(),
+            &fee_structure,
+            None,
+            &HashMap::new(),
+            &LoadedProgramsForTxBatch::default(),
+        );
+
+        assert_eq!(
+            result,
+            vec![(Err(TransactionError::InvalidWritableAccount), None)]
         );
     }
 }
